# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AssetLink Custody API
  description: |
    # AssetLink Custody API
    
    Secure Custody Infrastructure for Tokenized Real-World Assets (RWAs)
    
    AssetLink Custody is a custody-only backend system designed to securely manage 
    blockchain tokens that represent verified real-world assets.
    
    ## Authentication
    
    ### JWT Authentication (User Access)
    For dashboard and user operations, use JWT tokens:
    ```
    Authorization: Bearer YOUR_ACCESS_TOKEN
    ```
    
    ### API Key Authentication (Programmatic Access)
    For API integrations, use HMAC-SHA256 signature authentication:
    ```
    X-API-KEY: your_public_key
    X-SIGNATURE: hmac_signature
    X-TIMESTAMP: unix_timestamp
    X-USER-ID: end_user_identifier (optional for listing, required for mutations)
    ```
    
    **Two-Level Isolation:**
    - `X-API-KEY`: Identifies the platform owner (tenant)
    - `X-USER-ID`: Identifies the end user (issuer/investor)
    
    Platform owners see ALL activity. End users see only their own data.
    
    Signature payload: `METHOD + PATH + TIMESTAMP + BODY`
    
    ## Base URL
    - Development: `http://localhost:3000/v1`
    - Production: `https://api.assetlink.com/v1`
    
  version: 1.0.0
  contact:
    name: AssetLink Support
    email: support@assetlink.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/v1
    description: Development server
  - url: https://api.assetlink.com/v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for user authentication
    
    HmacAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      description: |
        HMAC-SHA256 signature authentication
        
        Required headers:
        - X-API-KEY: Public API key
        - X-SIGNATURE: HMAC signature
        - X-TIMESTAMP: Unix timestamp
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            statusCode:
              type: integer
            details:
              type: object
            timestamp:
              type: string
              format: date-time
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [CLIENT, ADMIN]
        status:
          type: string
          enum: [ACTIVE, SUSPENDED]
        createdAt:
          type: string
          format: date-time
    ApiKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        publicKey:
          type: string
        secretKey:
          type: string
          description: Only returned on creation
        userId:
          type: string
          format: uuid
        permissions:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  accessToken:
                    type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user and get access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  accessToken:
                    type: string
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get a new access token using refresh token cookie
      security: []
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Get authenticated user information
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/keys/my:
    get:
      tags:
        - Authentication
      summary: List my API keys
      description: Get all API keys for the authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKey'
    
    post:
      tags:
        - Authentication
      summary: Generate new API key
      description: Create a new API key for the authenticated user
      security:
        - BearerAuth: []
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
  /auth/keys/my/{id}:
    delete:
      tags:
        - Authentication
      summary: Revoke API key
      description: Revoke an API key
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: API key revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: API key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...